Refactor package structure to move ArgumentQueueImpl and ContextImpl to the impl package
