+++ src/main/kotlin/com/github/privatech/minimessage/impl/ArgumentQueueImpl.kt
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.github.privatech.minimessage.validation.argument
+package com.github.privatech.minimessage.impl
 
 import net.kyori.adventure.text.minimessage.Context
 import net.kyori.adventure.text.minimessage.tag.Tag
