--- src/main/kotlin/com/github/privatech/minimessage/validation/argument/Argument.kt
@@ -1,11 +1,0 @@
-package com.github.privatech.minimessage.validation.argument
-
-import net.kyori.adventure.text.minimessage.tag.Tag
-
-class Argument(val plainValue: String) : Tag.Argument {
-
-    override fun value(): String {
-        return plainValue
-    }
-
-}
